requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
bilibili-api-python>=17.0.0
python-dotenv==1.0.1
python-multipart==0.0.20
httpx==0.28.1
